deploy/requirements.txt
flask==2.0.2
pyjwt==1.7.1
SQLAlchemy==1.3.5
pandas==1.3.5
numpy==1.19.2
tensorflow==2.6.0
keras==2.6.0
scikit-learn==1.0.2
pickle5